Minor cleanings, not hardware-tested

+++ SMC.cpp
@@ -1,5 +1,6 @@
-#include "SMC.h"
 #include <Arduino.h>
+
+#include "SMC.h"
 
 SMCClass SMC;
 
@@ -26,7 +27,7 @@ int SMCClass::readByte()
 {
   char c;
   if(smcSerial.readBytes(&c, 1) == 0){ return -1; }
-  return (byte)c;
+  return static_cast<int>(c);
 }
 
 // required to allow motors to move
